fix(ink-veil): resolve FIX-001, FIX-002, FIX-003, FIX-004 from fca-review
Split 4 test files to comply with 3+12 rule (max 15 tests per file):
- regex-misc.test.ts (32→14/12/6)
- pipeline-system.e2e.test.ts (18→7/10)
- loader-priority.test.ts (17→10/7)
- dictionary.test.ts (17→8/9)
